[refactor] hypergan.samplers.viewer -> hypergan.viewer
